requirements.txt
psycopg2-binary==2.9.3
selenium==4.0.0
beautifulsoup4==4.10.0
pandas==1.4.0
requests==2.26.0
numpy==1.22.0